+++ database/postgres.go
@@ -54,7 +54,7 @@ func (c *postgresConnector) GetSchemas() ([]string, error) {
 	return schemas, nil
 }
 
-func (c *postgresConnector) GetTables(schemaNames []string) ([]TableNameResult, error) {
+func (c *postgresConnector) GetTables(schemaNames []string) ([]TableDetail, error) {
 	schemaSearch := "{" + strings.Join(schemaNames, ",") + "}"
 	rows, err := c.db.Query(`
 		select table_schema, table_name
@@ -66,9 +66,9 @@ func (c *postgresConnector) GetTables(schemaNames []string) ([]TableNameResult, 
 		return nil, err
 	}
 
-	var tables []TableNameResult
+	var tables []TableDetail
 	for rows.Next() {
-		var table TableNameResult
+		var table TableDetail
 		if err = rows.Scan(&table.Schema, &table.Name); err != nil {
 			return nil, err
 		}
@@ -81,7 +81,7 @@ func (c *postgresConnector) GetTables(schemaNames []string) ([]TableNameResult, 
 	return tables, nil
 }
 
-func (c *postgresConnector) GetColumns(tableName TableNameResult) ([]ColumnResult, error) {
+func (c *postgresConnector) GetColumns(tableName TableDetail) ([]ColumnResult, error) {
 	rows, err := c.db.Query(`
         select c.column_name,
                (case
@@ -105,14 +105,14 @@ func (c *postgresConnector) GetColumns(tableName TableNameResult) ([]ColumnResul
         from information_schema.columns c
                  left join pg_type typ on c.udt_name = typ.typname
                  left join pg_enum enu on typ.oid = enu.enumtypid
-        where c.table_name = $1
+        where c.table_name = $1  and c.table_schema = $2
         group by c.column_name,
                  c.table_name,
                  c.data_type,
                  c.udt_name,
                  c.ordinal_position
         order by c.ordinal_position;
-		`, tableName.Name)
+		`, tableName.Name, tableName.Schema)
 	if err != nil {
 		return nil, err
 	}
@@ -133,7 +133,7 @@ func (c *postgresConnector) GetColumns(tableName TableNameResult) ([]ColumnResul
 	return columns, nil
 }
 
-func (c *postgresConnector) GetConstraints(tableName TableNameResult) ([]ConstraintResult, error) {
+func (c *postgresConnector) GetConstraints(tableName TableDetail) ([]ConstraintResult, error) {
 	rows, err := c.db.Query(`
 	select fk.table_name,
 		   pk.table_name,
@@ -161,8 +161,8 @@ func (c *postgresConnector) GetConstraints(tableName TableNameResult) ([]Constra
 			 inner join information_schema.table_constraints fk on c.constraint_name = fk.constraint_name
 			 inner join information_schema.table_constraints pk on c.unique_constraint_name = pk.constraint_name
 			 inner join information_schema.key_column_usage kcu on c.constraint_name = kcu.constraint_name
-	where fk.table_name = $1 or pk.table_name = $1;
-		`, tableName.Name)
+	where c.constraint_schema = $1 and (fk.table_name = $2 or pk.table_name = $2);
+		`, tableName.Schema, tableName.Name)
 	if err != nil {
 		return nil, err
 	}