+++ database/table_name_util.go
@@ -5,20 +5,20 @@ import (
 	"strings"
 )
 
-func ParseTableName(value string, selectedSchemas []string) (TableNameResult, error) {
+func ParseTableName(value string, selectedSchemas []string) (TableDetail, error) {
 	parts := strings.Split(value, ".")
 
 	if len(parts) == 1 {
 		if len(selectedSchemas) != 1 {
-			return TableNameResult{}, errors.New("Could not parse table name")
+			return TableDetail{}, errors.New("If table names do not specify the schema, exactly one selected schema should be present")
 		}
 
-		return TableNameResult{Schema: selectedSchemas[0], Name: parts[0]}, nil
+		return TableDetail{Schema: selectedSchemas[0], Name: parts[0]}, nil
 	}
 
 	if len(parts) == 2 {
-		return TableNameResult{Schema: parts[0], Name: parts[1]}, nil
+		return TableDetail{Schema: parts[0], Name: parts[1]}, nil
 	}
 
-	return TableNameResult{}, errors.New("Could not parse table name")
+	return TableDetail{}, errors.New("Could not parse table name")
 }